add changes

+++ blog.php
@@ -17,10 +17,14 @@ if(array_key_exists('nazwa', $_GET) && file_exists("blogs/" . $_GET['nazwa'] . "
                 echo('<a href="'."blogs/".$_GET['nazwa']."/".$attachment.'" download>'.$attachment.'</a>');
             }
             echo("<br> komentarze:");
-            $comments = array_diff(scandir("blogs/".$_GET['nazwa']."/" . $post . ".k"), array('.', '..'));
-            foreach($comments as $comment) {
-                echo("<br>" . file_get_contents("blogs/".$_GET['nazwa']."/" . $post . ".k/" . $comment));
-            }   
+            if (file_exists("blogs/".$_GET['nazwa']."/" . $post . ".k")) {
+                $comments = array_diff(scandir("blogs/".$_GET['nazwa']."/" . $post . ".k"), array('.', '..'));
+                foreach($comments as $comment) {
+                    echo("<br>" . file_get_contents("blogs/".$_GET['nazwa']."/" . $post . ".k/" . $comment));
+                }
+            } else {
+                echo("Nie zamieszczono jeszcze żadnych komentarzy");
+            }
             echo("<br>");
             echo('<a href="./comment.php?postno='.$post."&blogname=".$_GET['nazwa'].'">Komentuj</a>');
             echo("</p>");
